Moved image_pipeline header files into include directory
Updated CMakeLists accordingly
