+++ intra_process_demo/include/image_pipeline/camera_node.hpp
