+++ intra_process_demo/include/image_pipeline/common.hpp
