+++ intra_process_demo/include/image_pipeline/image_view_node.hpp
