+++ intra_process_demo/include/image_pipeline/watermark_node.hpp
