+++ intra_process_demo/src/image_pipeline/image_pipeline_with_two_image_view.cpp
@@ -16,9 +16,9 @@
 
 #include <memory>
 
-#include "camera_node.hpp"
-#include "image_view_node.hpp"
-#include "watermark_node.hpp"
+#include "image_pipeline/camera_node.hpp"
+#include "image_pipeline/image_view_node.hpp"
+#include "image_pipeline/watermark_node.hpp"
 
 int main(int argc, char * argv[])
 {
