+++ intra_process_demo/src/image_pipeline/watermark_node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "watermark_node.hpp"
+#include "image_pipeline/watermark_node.hpp"
 
 #include <rclcpp/rclcpp.hpp>
